Da lhe amaral

+++ Assets/Scripts/IAJ.Unity/DecisionMaking/MCTS/MCTSNode.cs
@@ -35,7 +35,7 @@ namespace Assets.Scripts.IAJ.Unity.DecisionMaking.MCTS
             }
             toReturn += tabSpaces + " <N>" +(int)N +"</N>";
             toReturn += tabSpaces + " <Q>" + (int)Q + "</Q>";
-            toReturn += tabSpaces + " <Q_N_div>"+ (int)(Q / N) + "</Q_N_div>";
+            toReturn += tabSpaces + " <Q_N_div>"+ (Q / N) + "</Q_N_div>";
             //if (ChildNodes.Count > 0) {
             //    toReturn += tabSpaces + " <Number_Childs> " + ChildNodes.Count + " </Number_Childs>";
             //}
